Enable HashSetTest.
